+++ src/main/java/frc/robot/commands/test/CircleDrive.java
@@ -1,11 +1,8 @@
 package frc.robot.commands.test;
 
-import edu.wpi.first.math.MathUtil;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.kinematics.SwerveModuleState;
 import edu.wpi.first.util.WPIUtilJNI;
 import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.DrivetrainImpl;
+import frc.robot.subsystems.drive.DrivetrainImpl;
 
 /**
  * Drives robot in a circle. There is often drift due to inaccuracy.
