+++ src/main/java/frc/robot/controls/BaseDriverConfig.java
@@ -5,7 +5,7 @@ import edu.wpi.first.networktables.GenericEntry;
 import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import frc.robot.constants.miscConstants.OIConstants;
 import frc.robot.constants.swerve.DriveConstants;
-import frc.robot.subsystems.DrivetrainImpl;
+import frc.robot.subsystems.drive.DrivetrainImpl;
 import frc.robot.util.DynamicSlewRateLimiter;
 import frc.robot.util.MathUtils;
