+++ src/main/java/frc/robot/controls/Ex3DProDriverConfig.java
@@ -5,7 +5,7 @@ import edu.wpi.first.wpilibj2.command.InstantCommand;
 import frc.robot.commands.SetFormationX;
 import frc.robot.constants.miscConstants.OIConstants;
 import frc.robot.constants.swerve.DriveConstants;
-import frc.robot.subsystems.DrivetrainImpl;
+import frc.robot.subsystems.drive.DrivetrainImpl;
 import lib.controllers.Ex3DProController;
 import lib.controllers.Ex3DProController.Ex3DProAxis;
 import lib.controllers.Ex3DProController.Ex3DProButton;
