+++ src/main/java/frc/robot/subsystems/drive/ModuleSim.java
@@ -1,16 +1,15 @@
-package frc.robot.subsystems;
+package frc.robot.subsystems.drive;
 
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.kinematics.SwerveModulePosition;
 import edu.wpi.first.math.kinematics.SwerveModuleState;
-import edu.wpi.first.wpilibj.shuffleboard.ShuffleboardTab;
 import frc.robot.constants.Constants;
 import frc.robot.constants.swerve.ModuleConstants;
 
 /**
  * Swerve module for drivetrain to be used inside of simulation.
  *
- * @see frc.robot.subsystems.Module
+ * @see Module
  */
 public class ModuleSim extends Module {
